removing selenium tests, moving test/lint config locations, updating travis
