+++ iati/pytest.ini
@@ -1,0 +1,3 @@
+[pytest]
+DJANGO_SETTINGS_MODULE = iati.settings.dev
+python_files = tests.py test_*.py *_tests.py
