+++ iati/requirements_dev.txt
@@ -6,7 +6,6 @@ pytest-django==3.1.2
 splinter==0.10.0
 pytest-splinter==2.0.0
 responses==0.9.0
-selenium
 
 # Linters
 flake8==3.5.0
