--- iati/tests/data/activities.json
@@ -1,69 +1,0 @@
-{
-  "help": "https://iatiregistry.org/api/3/action/help_show?name=package_search",
-  "success": true,
-  "result": {
-    "count": 5699,
-    "sort": "title_string asc",
-    "facets": {
-      "extras_activity_count": {
-        "10122": 1,
-        "1012": 2,
-        "1011": 1,
-        "101": 5,
-        "1002": 1,
-        "1001": 1,
-        "100": 2,
-        "10": 107,
-        "1": 1096,
-        "0": 4
-      }
-    },
-    "results": [],
-    "search_facets": {
-      "extras_activity_count": {
-        "items": [{
-          "count": 1,
-          "display_name": "10122",
-          "name": "10122"
-        }, {
-          "count": 2,
-          "display_name": "1012",
-          "name": "1012"
-        }, {
-          "count": 1,
-          "display_name": "1011",
-          "name": "1011"
-        }, {
-          "count": 5,
-          "display_name": "101",
-          "name": "101"
-        }, {
-          "count": 1,
-          "display_name": "1002",
-          "name": "1002"
-        }, {
-          "count": 1,
-          "display_name": "1001",
-          "name": "1001"
-        }, {
-          "count": 2,
-          "display_name": "100",
-          "name": "100"
-        }, {
-          "count": 107,
-          "display_name": "10",
-          "name": "10"
-        }, {
-          "count": 1096,
-          "display_name": "1",
-          "name": "1"
-        }, {
-          "count": 4,
-          "display_name": "0",
-          "name": "0"
-        }],
-        "title": "extras_activity_count"
-      }
-    }
-  }
-}
