--- iati/tests/data/strings.py
@@ -1,47 +1,0 @@
-"""A module for different test string types."""
-
-SHORT_TEXT = 'Title or something'
-
-MEDIUM_TEXT = '''Lorem ipsum dolor sit amet, nisl quam, odio ut mauris, mi et nullam. Facilisis penatibus a, integer fusce vel, auctor elit. Aliquet viverra nunc, wisi a wisi. Suspendisse praesent, nec nisl, nam scelerisque sociis.'''
-
-LONG_TEXT = '''Lorem ipsum dolor sit amet, nulla magna ut elit arcu donec adipiscing, metus nunc vel suscipit posuere risus rutrum, mauris velit ultricies platea mi, integer adipiscing et id lectus rutrum. In aliquam nisl ante vehicula, massa quis quis maecenas vestibulum. Commodo inceptos quis vestibulum elit, et feugiat wisi, quam sollicitudin suscipit eget volutpat condimentum, eu morbi, lacus laoreet suspendisse phasellus. Dolor non mauris aenean curabitur ipsum augue, ante aliquam, auctor et aenean, feugiat odio elit est facilisis etiam, nullam wisi in sed wisi. Nec tortor vel, arcu et congue. Velit risus consectetuer incididunt id arcu, integer at sem vitae volutpat aliquet, arcu wisi non integer lacus sit et, in velit in purus. Dolor sagittis quam magna, at proin neque consectetuer. Et aptent consequat metus ante lectus faucibus, mi eu vulputate dignissim, pede et viverra et morbi sit praesent, velit nulla justo fringilla cras id at, tellus condimentum sit nec. Nulla ullamcorper dictum justo magna nibh arcu, vitae etiam sit fermentum turpis.\nOrci habitasse cras sem, vitae blandit tincidunt pede quam. Morbi elementum egestas id, sollicitudin molestias ultricies orci nonummy pede, tellus nisl ut magna, sed mattis hendrerit eget, pede non enim. Nam cubilia dictumst libero nonummy accumsan, lacus arcu auctor. Condimentum tellus. Mi nulla tincidunt nisl adipiscing elit. Nam et malesuada praesent fringilla amet, suscipit montes, eget libero litora, orci hymenaeos ullamcorper. Vulputate luctus magna eu vivamus, quisque sed tempor eget ipsum ut, in justo molestie ligula massa, pharetra duis. Congue elit ultricies, rutrum sapien vitae, eget sed, in lorem. Nulla sit purus etiam lorem rhoncus tempus, aliquam sapien, sed pellentesque facilisis mauris, et at mi consequat ultrices egestas, semper proin massa duis. Vel duis lorem donec elit purus augue. Iaculis lectus mauris fringilla quam ligula, interdum donec ante molestie ultrices mauris mauris.\nTellus scelerisque sit fusce, ac leo vitae nunc sem viverra. Ipsum elementum non pellentesque quam, amet libero. Parturient mus mauris in donec ad, eros dictum leo felis, lectus wisi praesent enim semper iaculis consequat, at sollicitudin nunc sed odit, nibh elementum convallis. Sed cursus. Nulla risus, sem vel. Pulvinar sit vestibulum, ac urna eleifend morbi nec pulvinar, morbi cras aliquam molestie, ut quis pharetra mi lacus ullamcorper, condimentum scelerisque ullamcorper orci. Quis faucibus fusce aliquam arcu proin aliquet, eget nec dolor quis at quam, tellus nam vel bibendum erat facilisi. Massa tellus, duis dolores, dui magna etiam integer dolor interdum, suscipit fringilla sit id risus quis tellus. Vivamus ac, lorem auctor iaculis id ultrices lacus, dictum diam nonummy nec mi dignissim. Volutpat nisl metus felis, risus tempus enim lacus sed sapien ac, tristique sed sed cras, consequat dolor lectus vitae lorem torquent, quis urna sociis arcu class.\nSit wisi in, a praesent eu mauris aliquam purus pellentesque, tellus egestas sem egestas vivamus a, elementum lorem a. Leo eget et risus. Malesuada vivamus sociis vestibulum orci, fames urna, ipsum felis aliquam. Integer cursus vel sed, pellentesque malesuada tincidunt nullam. Adipiscing venenatis netus auctor semper sem.'''  # noqa: E501
-
-RAW_HTML = '''
-<p align="justify">
-Lorem ipsum dolor sit amet, accumsan neque lectus ligula mattis netus. A pede posuere. Sem cras torquent, mi lorem augue, risus amet et semper vehicula est. Ipsum vivamus, morbi tincidunt, suspendisse non justo montes, sociis nullam elit metus interdum integer, donec amet nulla consequat praesent. Integer mauris morbi quis sapien, pretium aenean pulvinar diam iaculis eros volutpat, amet turpis non placerat suscipit phasellus lacus, lectus sed sem libero aliquam ligula. Consequat tincidunt vitae proin maecenas habitant lorem, parturient lacinia vehicula a dolor, maecenas lectus interdum arcu lobortis quisque mauris, praesent quis primis, purus aliquam ipsum tristique. Curabitur nec, erat tempora, accumsan et enim consectetuer magnis ligula, justo fermentum lacus imperdiet, eu sed habitasse.'''  # noqa: E501
-RAW_HTML += '''
-</p>
-<p>
-    <ul>
-        <li>item 1</li>
-        <li>item 2</li>
-        <li>item 3</li>
-        <li>item 4</li>
-    </ul>
-</p>
-<table>
-    <tbody>
-        <tr>
-            <td>this</td>
-            <td>is</td>
-            <td>a</td>
-            <td>little</td>
-            <td>table</td>
-        </tr>
-        <tr>
-            <td>0</td>
-            <td>0</td>
-            <td>0</td>
-            <td>0</td>
-            <td>0</td>
-        </tr>
-        <tr>
-            <td>1</td>
-            <td>1</td>
-            <td>1</td>
-            <td>1</td>
-            <td>1</td>
-        </tr>
-    </tbody>
-</table>
-'''
