--- iati/tests/support_functional_tests.py
@@ -1,1 +1,0 @@
-"""A module of functional tests for the support page and its sub pages."""
