--- pytest.ini
@@ -1,3 +1,0 @@
-[pytest]
-DJANGO_SETTINGS_MODULE = iati.settings.dev
-python_files = tests.py test_*.py *_tests.py
