--- setup.cfg
@@ -1,10 +1,0 @@
-[flake8]
-exclude = migrations  # migrations should not be flake8 linted since they are automatically generated
-max-line-length = 255
-ignore = F841
-
-[pydocstyle]
-add_ignore = D203,D213
-add_select = D212,D404
-convention = pep257
-match = (?!test_)(?!\d{4}_).*\.py  # Ignores most migrations files, though full fix pending resolution of https://github.com/PyCQA/pydocstyle/issues/175
